create external stylesheet to clean  up components
