+++ Screens/SignIn.jsx
@@ -19,6 +19,7 @@ import {
 } from "../utils/auth";
 import * as WebBrowser from "expo-web-browser";
 import { createUserAccount, getUserAccount } from "../utils/dbApi";
+import { signInStyles } from '../Styles/styles.js'
 
 WebBrowser.maybeCompleteAuthSession();
 
@@ -74,141 +75,52 @@ function SignIn() {
   return (
     <KeyboardAvoidingView
       behavior={Platform.OS === "*" ? "padding" : "height"}
-      style={styles.container}
+      style={signInStyles.container}
     >
-      <Text style={styles.heading}>Welcome</Text>
+      <Text style={signInStyles.heading}>Welcome</Text>
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-        <View style={styles.inputContainer}>
+        <View style={signInStyles.inputContainer}>
           <TextInput
             placeholder="email"
             value={email}
             onChangeText={(text) => setEmail(text)}
-            style={styles.Input}
+            style={signInStyles.Input}
           />
           <TextInput
             placeholder="password"
             value={password}
             onChangeText={(text) => setPassword(text)}
-            style={styles.Input}
+            style={signInStyles.Input}
             secureTextEntry
           />
         </View>
       </TouchableWithoutFeedback>
-      <View style={styles.buttonContainer}>
+      <View style={signInStyles.buttonContainer}>
         <TouchableOpacity
           onPress={handleSignIn}
-          style={[styles.button, styles.buttonOutline]}
+          style={[signInStyles.button, signInStyles.buttonOutline]}
         >
-          <Text style={styles.buttonOutlineText}>Sign in</Text>
+          <Text style={signInStyles.buttonOutlineText}>Sign in</Text>
         </TouchableOpacity>
         <TouchableOpacity disabled={!request} onPress={() => promptAsync()}>
-          <View style={styles.GoogleButton}>
+          <View style={signInStyles.GoogleButton}>
             <Image
-              style={styles.GoogleImage}
+              style={signInStyles.GoogleImage}
               source={require("../assets/google-img.png")}
             />
-            <Text style={styles.GoogleButtonText}>Sign in with Google</Text>
+            <Text style={signInStyles.GoogleButtonText}>Sign in with Google</Text>
           </View>
         </TouchableOpacity>
       </View>
-      <View style={styles.pageLink}>
+      <View style={signInStyles.pageLink}>
         <Text>Don't have an account</Text>
         <TouchableOpacity onPress={() => navigation.navigate("Sign Up")}>
-          <Text style={styles.signUp}> Sign Up</Text>
+          <Text style={signInStyles.signUp}> Sign Up</Text>
         </TouchableOpacity>
       </View>
     </KeyboardAvoidingView>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    fontSize: 53,
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "white",
-    // width: Dimensions.get("window"),
-    // height: Dimensions.get("window"),
-  },
-  heading: { color: "black", fontSize: 40, margin: 20 },
-  user: {
-    justifyContent: "center",
-    textAlign: "center",
-  },
-  Input: {
-    backgroundColor: "#f2f2f2",
-    paddingHorizontal: 15,
-    paddingVertical: 10,
-    borderRadius: 15,
-    marginTop: 25,
-    // padding: 55,
-  },
-  inputContainer: {
-    width: "80%",
-  },
-  button: {
-    alignItems: "center",
-    backgroundColor: "white",
-    width: "100%",
-    padding: 10,
-    borderRadius: 15,
-  },
-  buttonContainer: {
-    width: "60%",
-    justifyContent: "center",
-    alignContent: "center",
-    marginTop: 30,
-  },
-  buttonOutline: {
-    padding: 15,
-    borderWidth: 2,
-    borderColor: "white",
-    backgroundColor: "#24a0ed",
-    marginBottom: 30,
-  },
-  buttonOutlineText: {
-    color: "white",
-    fontSize: 18,
-    fontWeight: "bold",
-  },
-  pageLink: {
-    flexDirection: "row",
-    alignItems: "flex-end",
-    justifyContent: "center",
-    marginTop: 40,
-    marginBottom: 20,
-  },
-  signUp: {
-    color: "#24a0ed",
-    fontWeight: "bold",
-  },
-  twoButtons: {
-    flexDirection: "row",
-    justifyContent: "space-around",
-  },
-  GoogleButton: {
-    padding: 15,
-    borderWidth: 1,
-    borderRadius: 15,
-    borderColor: "black",
-    justifyContent: "center",
-    alignitems: "center",
-    flexDirection: "row",
-  },
-  GoogleImage: {
-    width: 20,
-    height: 20,
-    marginLeft: 5,
-    alignContent: "center",
-    justifyContent: "center",
-  },
-  GoogleButtonText: {
-    fontSize: 18,
-    fontWeight: "bold",
-    color: "grey",
-    marginLeft: 10,
-  },
-});
 
 export default SignIn;
